+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v0.1.4 (2021-02-01)
+* Fix isinstance bug
+
+# v0.1.3 (2021-02-01)
+* Fix typing.Any --> typing.Union
+
+# v0.1.2 (2021-01-29)
+* Add stripe_customer_id to app_metadata fields
+* Add typing to user_manager
+
 # v0.1.1 (2020-07-27)
 
 * Add get_logs script